# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_models.ServerAccount:
    properties:
      backup_dir_info:
        $ref: '#/definitions/internal_models.ServerAccountBackupDirInfo'
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      gid:
        type: integer
      host:
        type: string
      name:
        type: string
      not_exists_in_server:
        type: boolean
      port:
        type: integer
      pwd:
        type: string
      uid:
        type: integer
      updated_at:
        type: string
    type: object
  internal_models.ServerAccountBackupDirInfo:
    properties:
      backup_dir:
        type: string
      dir_exists:
        type: boolean
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      output:
        type: string
      path_exists:
        type: boolean
    type: object
  internal_models.ServerAccountBackupDirResponse:
    properties:
      backup_dir:
        type: string
      dir_exists:
        type: boolean
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      output:
        type: string
      path_exists:
        type: boolean
    type: object
  internal_models.ServerAccountCreateRequest:
    properties:
      account_name:
        type: string
      account_pwd:
        type: string
      host:
        type: string
      port:
        type: integer
    type: object
  internal_models.ServerAccountCreateResponse:
    type: object
  internal_models.ServerAccountDeleteRequest:
    properties:
      account_name:
        type: string
      backup:
        type: boolean
      host:
        type: string
      port:
        type: integer
    type: object
  internal_models.ServerAccountDeleteResponse:
    properties:
      backup_dir:
        type: string
    type: object
  internal_models.ServerAccountInfos:
    properties:
      accounts:
        items:
          $ref: '#/definitions/internal_models.ServerAccount'
        type: array
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      output:
        type: string
    type: object
  internal_models.ServerAccountUpdateRequest:
    properties:
      account_name:
        type: string
      account_pwd:
        type: string
      host:
        type: string
      port:
        type: integer
      recover:
        description: Recover 该账户是从删除的账户中恢复
        type: boolean
      recover_backup:
        description: RecoverBackup 指定是否要恢复backup的用户目录文件夹。
        type: boolean
    type: object
  internal_models.ServerAccountUpdateResponse:
    type: object
  internal_models.ServerBasic:
    properties:
      admin_account_name:
        type: string
      admin_account_pwd:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      host:
        type: string
      os_type:
        type: string
      port:
        type: integer
      updated_at:
        type: string
    type: object
  internal_models.ServerCPUHardwareInfo:
    properties:
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      info:
        $ref: '#/definitions/internal_models.ServerCPUs'
      output:
        type: string
    type: object
  internal_models.ServerCPUMemProcessesUsageInfo:
    properties:
      cpu_mem_usage:
        $ref: '#/definitions/internal_models.ServerCPUMemUsage'
        description: CPUMemUsage 服务器总的CPU，内存使用率。
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      output:
        type: string
      process_infos:
        description: ProcessInfos 全部进程信息。
        items:
          $ref: '#/definitions/internal_models.ServerProcessInfo'
        type: array
    type: object
  internal_models.ServerCPUMemUsage:
    properties:
      mem_total:
        description: MemTotal 内存总量，使用字符串固定死
        type: string
      mem_usage:
        description: MemUsage 总内存使用（比例：如3600MB/8000MB）
        type: number
      user_cpu_usage:
        description: UserProcCPUUsage 记录用户进程的CPU使用率。（总比例）
        type: number
    type: object
  internal_models.ServerCPUs:
    properties:
      architecture:
        description: Architecture 架构
        type: string
      cores:
        description: Cores CPU核数
        type: integer
      model_name:
        description: ModelName 如：Intel(R) Xeon(R) CPU E5-2682 v4 @ 2.50GHz
        type: string
      threads_per_core:
        description: ThreadsPerCore 每个核心可以跑几个线程
        type: integer
    type: object
  internal_models.ServerConnectionTestResponse:
    properties:
      cause:
        type: string
      connected:
        type: boolean
    type: object
  internal_models.ServerCreateRequest:
    properties:
      admin_account_name:
        type: string
      admin_account_pwd:
        type: string
      host:
        type: string
      os_type:
        type: string
      port:
        type: integer
    type: object
  internal_models.ServerCreateResponse:
    type: object
  internal_models.ServerDeleteRequest:
    properties:
      host:
        type: string
      port:
        type: integer
    type: object
  internal_models.ServerDeleteResponse:
    type: object
  internal_models.ServerGPU:
    properties:
      product:
        description: Product 产品名。
        type: string
    type: object
  internal_models.ServerGPUHardwareInfos:
    properties:
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      infos:
        items:
          $ref: '#/definitions/internal_models.ServerGPU'
        type: array
      output:
        type: string
    type: object
  internal_models.ServerGPUUsageInfo:
    properties:
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      output:
        type: string
    type: object
  internal_models.ServerHardwareInfo:
    properties:
      cpu_hardware_info:
        $ref: '#/definitions/internal_models.ServerCPUHardwareInfo'
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      gpu_hardware_infos:
        $ref: '#/definitions/internal_models.ServerGPUHardwareInfos'
      output:
        type: string
    type: object
  internal_models.ServerInfo:
    properties:
      access_failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
        description: AccessFailedInfo 指定了当该服务器连接失败时的信息。如果该字段不为空，那么其他字段才有意义。
      account_infos:
        $ref: '#/definitions/internal_models.ServerAccountInfos'
        description: AccountInfos 记录服务器账户信息。
      basic:
        $ref: '#/definitions/internal_models.ServerBasic'
        description: Basic 基本的Server目录信息
      cpu_mem_processes_usage_info:
        $ref: '#/definitions/internal_models.ServerCPUMemProcessesUsageInfo'
        description: CPUMemProcessesUsageInfo CPU，内存，进程的使用资源信息。（Top指令）
      hardware_info:
        $ref: '#/definitions/internal_models.ServerHardwareInfo'
        description: ServerHardwareInfo 硬件元信息
      remote_accessing_usage_info:
        $ref: '#/definitions/internal_models.ServerRemoteAccessingUsagesInfo'
        description: RemoteAccessingUsageInfo 正在从远端访问的用户的使用信息
      server_gpu_usage_info:
        $ref: '#/definitions/internal_models.ServerGPUUsageInfo'
        description: GPUUsageInfo 当前该Server总的GPU利用率信息。（当前为string，具体待定）
    type: object
  internal_models.ServerInfoLoadingFailedInfo:
    properties:
      cause_description:
        description: 描述具体原因。
        type: string
    type: object
  internal_models.ServerInfoResponse:
    properties:
      access_failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
        description: AccessFailedInfo 指定了当该服务器连接失败时的信息。如果该字段不为空，那么其他字段才有意义。
      account_infos:
        $ref: '#/definitions/internal_models.ServerAccountInfos'
        description: AccountInfos 记录服务器账户信息。
      basic:
        $ref: '#/definitions/internal_models.ServerBasic'
        description: Basic 基本的Server目录信息
      cpu_mem_processes_usage_info:
        $ref: '#/definitions/internal_models.ServerCPUMemProcessesUsageInfo'
        description: CPUMemProcessesUsageInfo CPU，内存，进程的使用资源信息。（Top指令）
      hardware_info:
        $ref: '#/definitions/internal_models.ServerHardwareInfo'
        description: ServerHardwareInfo 硬件元信息
      remote_accessing_usage_info:
        $ref: '#/definitions/internal_models.ServerRemoteAccessingUsagesInfo'
        description: RemoteAccessingUsageInfo 正在从远端访问的用户的使用信息
      server_gpu_usage_info:
        $ref: '#/definitions/internal_models.ServerGPUUsageInfo'
        description: GPUUsageInfo 当前该Server总的GPU利用率信息。（当前为string，具体待定）
    type: object
  internal_models.ServerInfosResponse:
    properties:
      infos:
        items:
          $ref: '#/definitions/internal_models.ServerInfo'
        type: array
      total_count:
        type: integer
    type: object
  internal_models.ServerProcessInfo:
    properties:
      command:
        description: Command 命令
        type: string
      cpu_usage:
        description: CPU利用率。
        type: number
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      gpu_usage:
        description: GPU利用率（不一定能查到）
        type: string
      mem_usage:
        description: 内存利用率
        type: number
      output:
        type: string
      owner_account_name:
        description: OwnerAccountName 该进程由哪个用户启动。
        type: string
      pid:
        description: PID 进程号。
        type: integer
    type: object
  internal_models.ServerRemoteAccessingAccountInfo:
    properties:
      account_name:
        type: string
      what:
        description: What 表示该远程访问的用户正在执行的命令。
        type: string
    type: object
  internal_models.ServerRemoteAccessingUsagesInfo:
    properties:
      failed_info:
        $ref: '#/definitions/internal_models.ServerInfoLoadingFailedInfo'
      infos:
        items:
          $ref: '#/definitions/internal_models.ServerRemoteAccessingAccountInfo'
        type: array
      output:
        type: string
    type: object
  internal_models.SessionsCheckResponse:
    properties:
      userID:
        type: integer
    type: object
  internal_models.SessionsCreateRequest:
    properties:
      name:
        type: string
      pwd:
        type: string
    type: object
  internal_models.SessionsCreateResponse:
    properties:
      token:
        type: string
    type: object
  internal_models.SessionsDestroyRequest:
    type: object
  internal_models.SessionsDestroyResponse:
    type: object
  internal_models.User:
    properties:
      admin:
        type: boolean
      created_at:
        type: integer
      id:
        type: integer
      name:
        type: string
      pwd:
        type: string
      updated_at:
        type: integer
    type: object
  internal_models.UsersCreateRequest:
    properties:
      name:
        type: string
      pwd:
        type: string
    type: object
  internal_models.UsersCreateResponse:
    properties:
      token:
        type: string
    type: object
  internal_models.UsersInfoResponse:
    properties:
      admin:
        type: boolean
      created_at:
        type: integer
      id:
        type: integer
      name:
        type: string
      pwd:
        type: string
      updated_at:
        type: integer
    type: object
  internal_models.UsersInfosResponse:
    properties:
      infos:
        items:
          $ref: '#/definitions/internal_models.User'
        type: array
      total_count:
        type: integer
    type: object
  internal_models.UsersUpdateRequest:
    properties:
      admin:
        type: boolean
      name:
        type: string
      pwd:
        type: string
    type: object
  internal_models.UsersUpdateResponse:
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a ServerServing API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ServerServing Web API
  version: "1.0"
paths:
  /api/v1/servers/:
    delete:
      parameters:
      - description: serverDeleteRequest
        in: body
        name: serverDeleteRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.ServerDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerDeleteResponse'
      summary: 删除server。
      tags:
      - server
    get:
      parameters:
      - in: query
        name: from
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: size
        type: integer
      - description: WithAccounts 加载账户信息的参数，为nil则不加载
        in: query
        name: with_accounts
        type: boolean
      - description: WithBackupDirInfo 指定是否加载用户备份文件夹的信息。
        in: query
        name: with_backup_dir_info
        type: boolean
      - description: WithCPUMemProcessesUsage 指定是否加载CPU，内存，进程的使用信息。
        in: query
        name: with_cmp_usages
        type: boolean
      - description: WithGPUUsages 指定是否加载GPU的使用信息。
        in: query
        name: with_gpu_usages
        type: boolean
      - description: WithHardwareInfo 指定是否加载硬件的元信息
        in: query
        name: with_hardware_info
        type: boolean
      - description: WithRemoteAccessUsages 指定是否加载正在远程登录这台服务器的用户信息。
        in: query
        name: with_remote_access_usages
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerInfosResponse'
      summary: 查询多个server信息。
      tags:
      - server
    post:
      parameters:
      - description: serverCreateRequest
        in: body
        name: serverCreateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.ServerCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerCreateResponse'
      summary: 创建server。
      tags:
      - server
  /api/v1/servers/{host}/{port}:
    get:
      parameters:
      - description: host
        in: path
        name: host
        required: true
        type: string
      - description: port
        in: path
        name: port
        required: true
        type: integer
      - description: WithAccounts 加载账户信息的参数，为nil则不加载
        in: query
        name: with_accounts
        type: boolean
      - description: WithBackupDirInfo 指定是否加载用户备份文件夹的信息。
        in: query
        name: with_backup_dir_info
        type: boolean
      - description: WithCPUMemProcessesUsage 指定是否加载CPU，内存，进程的使用信息。
        in: query
        name: with_cmp_usages
        type: boolean
      - description: WithGPUUsages 指定是否加载GPU的使用信息。
        in: query
        name: with_gpu_usages
        type: boolean
      - description: WithHardwareInfo 指定是否加载硬件的元信息
        in: query
        name: with_hardware_info
        type: boolean
      - description: WithRemoteAccessUsages 指定是否加载正在远程登录这台服务器的用户信息。
        in: query
        name: with_remote_access_usages
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerInfoResponse'
      summary: 查询server信息。
      tags:
      - server
  /api/v1/servers/accounts:
    delete:
      parameters:
      - description: serverAccountDeleteRequest
        in: body
        name: serverAccountDeleteRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.ServerAccountDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerAccountDeleteResponse'
      summary: 删除一个服务器的账号。
      tags:
      - server_account
    post:
      parameters:
      - description: serverAccountCreateRequest
        in: body
        name: serverAccountCreateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.ServerAccountCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerAccountCreateResponse'
      summary: 创建一个服务器的sudo账号。
      tags:
      - server_account
    put:
      parameters:
      - description: serverAccountUpdateRequest
        in: body
        name: serverAccountUpdateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.ServerAccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerAccountUpdateResponse'
      summary: 更新，恢复一个服务器的账号。
      tags:
      - server_account
  /api/v1/servers/accounts/backupDir:
    get:
      parameters:
      - in: query
        name: account_name
        type: string
      - in: query
        name: host
        type: string
      - in: query
        name: port
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerAccountBackupDirResponse'
      summary: 获取一个账户的backup文件夹的相关信息
      tags:
      - server_account
  /api/v1/servers/connections/{host}/{port}:
    get:
      parameters:
      - description: host
        in: path
        name: host
        required: true
        type: string
      - description: port
        in: path
        name: port
        required: true
        type: integer
      - in: query
        name: account_name
        type: string
      - in: query
        name: account_pwd
        type: string
      - in: query
        name: os_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.ServerConnectionTestResponse'
      summary: 测试连通性
      tags:
      - server
  /api/v1/sessions/:
    delete:
      parameters:
      - description: destroyRequest
        in: body
        name: sessionsDestroyRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.SessionsDestroyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.SessionsDestroyResponse'
      summary: 退出session。（退出登录）
      tags:
      - session
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.SessionsCheckResponse'
      summary: 检查登录状态。
      tags:
      - session
    post:
      parameters:
      - description: createRequest
        in: body
        name: sessionsCreateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.SessionsCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.SessionsCreateResponse'
      summary: 创建session。（登录）
      tags:
      - session
  /api/v1/test/error_handler:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: test error handler
      tags:
      - test
  /api/v1/test/ping:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: ping
      tags:
      - test
  /api/v1/users/:
    get:
      parameters:
      - in: query
        name: from
        type: integer
      - in: query
        name: searchKeyword
        type: string
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.UsersInfosResponse'
      summary: 获取多个用户信息，可以添加关键字对姓名搜索。
      tags:
      - user
    post:
      parameters:
      - description: userCreateRequest
        in: body
        name: userCreateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.UsersCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.UsersCreateResponse'
      summary: 注册用户
      tags:
      - user
  /api/v1/users/{id}:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: x-token
        in: header
        name: x-token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.UsersInfoResponse'
      summary: 获取单个用户信息
      tags:
      - user
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: updateRequest
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/internal_models.UsersUpdateRequest'
      - description: x-token
        in: header
        name: x-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_models.UsersUpdateResponse'
      summary: 修改用户信息
      tags:
      - user
swagger: "2.0"